Refactor: Remove Ollama-specific config injection
I removed the `injectOllamaConfig` feature based on your feedback.

- I removed the `injectOllamaConfig` function from `crewai-web-ui/src/app/api/generate/script.utils.ts`.
- I removed the corresponding import and call to this function from `crewai-web-ui/src/app/api/generate/route.ts`.

The `/api/generate` route will no longer inject Ollama-specific llm configuration into generated scripts.
